==> Capstone/0_Pickle.ipynb <==
# Capstone Project: Udacity Machine Learning Engineer
## Hans Fricke 2018
### Transformation to python data

# import packages
import pandas as pd
import os

# display formatting
%matplotlib inline 
pd.options.display.float_format = '{:20,.2f}'.format

# change working directory
os.chdir("/Users/hansfricke/Dropbox/ND ML Engineer/Capstone")

#### Proposal data

headers = ['id', 
            'teacher_id',
            'teacher_prefix',
            'state',
            'datetime',
            'grade_category',
            'subject_categories',
            'subject_subcategories',
            'title',
            'essay_1',
            'essay_2',
            'essay_3',
            'essay_4',
            'resource_summary',
            'previously_posted_projects',
            'approved']
dtypes = { 'id' : 'str', 
           'teacher_id' : 'str',
           'teacher_prefix' : 'str',
           'state' : 'str',
           'datetime' : 'str',
           'grade_category' : 'str',
           'subject_categories' : 'str',
           'subject_subcategories' : 'str',
           'title' : 'str',
           'essay_1' : 'str',
           'essay_2' : 'str',
           'essay_3' : 'str',
           'essay_4' : 'str',
           'resource_summary' :  'str',
           'previously_posted_projects' : 'int',
           'approved' : 'int'}
parse_dates = ['datetime']
proposal_data = pd.read_csv('train.csv', header = 0, na_values = 'NaN',
                    names=headers, dtype=dtypes, parse_dates=parse_dates)

proposal_data.head()

proposal_data.describe()

proposal_data.to_pickle("Data/proposals.pkl")

#### Resources data

dtypes = { 'id': 'str', 
           'description': 'str',
           'quantity': 'int',
           'price': 'float'}
resources_data = pd.read_csv('resources.csv', na_values = 'NaN', dtype=dtypes)

resources_data.head()

resources_data.describe()

resources_data.to_pickle('Data/resources.pkl')
