==> Capstone/1_ExploratoryAnalysis_DataCleaning.ipynb <==
# Capstone Project: Udacity Machine Learning Engineer
## Hans Fricke 2018
### Data cleaning

# import packages
import pandas as pd
import os
import string
import datetime
import hunspell
hobj = hunspell.HunSpell('/Library/Spelling/en_US.dic',
                         '/Library/Spelling/en_US.aff')

from nltk.corpus import stopwords
stop_words = set(stopwords.words('english'))

# display formatting
%matplotlib inline 
pd.options.display.float_format = '{:20,.2f}'.format

# change working directory
os.chdir("/Users/hansfricke/Dropbox/ND ML Engineer/Capstone/Data")

#### Proposal data

proposals = pd.read_pickle('proposals.pkl')

proposals.head()

proposals['approved'].describe()

proposals['previously_posted_projects'].describe()

##### Data cleaning steps

1. create dummy_variables for categorical responses

##### 1) Dummy variables

# teacher prefix
proposals['teacher_prefix'].value_counts()/len(proposals['teacher_prefix'])

# state
proposals['state'].value_counts()/len(proposals['state'])

# grades
proposals['grade_category'].value_counts()/len(proposals['state'])

# main subjects
print(proposals['subject_categories'].value_counts())
print(len(proposals['subject_categories'].value_counts()))

# sub subjects
print(proposals['subject_subcategories'].value_counts())
print(len(proposals['subject_subcategories'].value_counts()))

- teacher prefix, state and grade can simply be transformed with get_dummies().
- the subject variables should be split at the comma and a dummy created for each subject.

# write a program to get list of all unique subjects
def get_subjects(series):
    subject_combinations = list(set(series))
    subjects = []
    for combination in subject_combinations:
        subjects.extend(combination.split(','))
        subjects = list(set(subject.replace(' ', '') for subject in subjects))
    return subjects    

subjects = get_subjects(proposals['subject_categories'])
subsubjects = get_subjects(proposals['subject_subcategories'])

print(subjects)
print(len(subjects))

print(subsubjects)
print(len(subsubjects))

# function to create a dummy if subject is included in subject combination
def create_subject_dummies(string, subject_list):
    subjects = string.split(',')
    subjects = list(set(subject.replace(' ', '') for subject in subjects))
    dummy_values = [int(subject in subjects) for subject in subject_list]
    return dummy_values        

subject_columns = proposals['subject_categories'].apply(lambda x: create_subject_dummies(x, subject_list=subjects))
proposals[subjects]=pd.DataFrame(subject_columns.tolist())

subsubject_columns = proposals['subject_subcategories'].apply(lambda x: create_subject_dummies(x, subject_list=subsubjects))
proposals[subsubjects]=pd.DataFrame(subsubject_columns.tolist())

proposals[subjects].mean().sort_values()

proposals[subsubjects].mean().sort_values()

proposals = proposals.join(pd.get_dummies(proposals[['teacher_prefix', 'state', 'grade_category']]))

#### Resources

resources = pd.read_pickle('resources.pkl')

resources.head()

resources.describe()

##### Steps

1. Calculate total price per item
2. Calculate total price for requested resources
3. Count number of items
4. Concatenate all items description
5. Keep one record per ID


###### 1. Calculate total price per item

resources['total_item_price'] = resources['price'] * resources['quantity']

##### 2. Calculate total price for requested resources

resources['total_price'] = resources.groupby('id')['total_item_price'].transform('sum')

###### 3. Count number of items

resources['num_items'] = resources.groupby('id')['id'].transform('count')

###### 4. Concatenate all items description

# fuction takes a series and joins all values to one string
# if any value not string, transform all values to string
# try/except fro computation speed
def join_items(series):
    try:
        joined_str = ' '.join(series)
    except TypeError:
        joined_str = ' '.join([str(x) for x in series])
    return joined_str

resources['total_description'] = resources.groupby('id')['description'].transform(join_items) 

resources.head()

##### 5. Keep one record per ID

new_resources = resources.drop_duplicates('id')

# Keep only new columns with aggregate information
new_resources = new_resources[['id', 'total_price', 'num_items', 'total_description']]

new_resources.describe()

#### Merge both data sets

pre_cleaned_proposals = proposals.merge(new_resources, on='id', how = 'left')

pre_cleaned_proposals.head()

#### Text cleaning

1. All lower case letters
2. Remove numbers and punctuation
3. Correct spelling mistakes
4. Stem words

# function to remove numbers and punctuation
def remove_punct_num(text):
    punct = str.maketrans(dict.fromkeys(string.punctuation))
    text = text.translate(punct)
    digits = str.maketrans(dict.fromkeys(string.digits))
    text = text.translate(digits)
    return text

# function to correct spelling
def correct_spelling(word):
    if not hobj.spell(word):
        if word in words_w_apostrophy.keys():
            word = words_w_apostrophy[word]
        else:
            correct_words = hobj.suggest(word)
            try:
                word = correct_words[0].replace("-", " ")
            except IndexError:
                # No correct suggestion for word, remove 
                word = ''
    return word

# dictionary with words that spell check get's wrong, words with appostrophe
words_w_apostrophy = { 'didnt': "didn't",
                  'dont': "do not",
                  'wasnt': "was not",
                  'arent': "are not",
                  'werent': "were not",
                  'doesnt': "does not",
                  'youve': "you have",
                  'whod': "who would",
                  'ive': "i have",
                  'wouldve': "would have",
                  'im': "i am",
                  'weve': "we have",
                  'thats': "that is",
                  'isnt': "is not",
                  'theyve': "they have",
                  'theyre': "they are",
                  'hasnt': "has not"}

# fuction that stems word
def stem(word):
    word = hobj.stem(word)
    try:
        return word[0].decode("utf-8") 
    except IndexError: #Not real word, no stem available
        return ''

def process_text(text):
    # remove punctuation and numbers
    text = remove_punct_num(text)
    # set everything to lower case
    text = text.lower()
    # split the string into words for individual processing
    words = text.split()
    # new list for processed words
    processed = []
    for word in words:
        # correct spelling
        word = correct_spelling(word)
        # somtimes corrected word contains two new words
        # seperated here and porcessed individually
        split_words = word.split()
        for split_word in split_words:
            # remove stop words
            if split_word not in stop_words:
                # stem the word and append to processed list
                processed.append(stem(split_word))
    # return a processed string            
    return ' '.join(processed)

I combine essay questions 1 and 2 as well as 3 and 4 because the number of questions changed after May 17, 2016.
Before May 17, 2016, the essay questions were:
    
* project_essay_1: "Introduce us to your classroom"
* project_essay_2: "Tell us more about your students"
* project_essay_3: "Describe how your students will use the materials you're requesting"
* project_essay_4: "Close by sharing why your project will make a difference"

Starting on May 17, 2016, the number of essays was reduced from 4 to 2, and the prompts for the first 2 essays were changed to the following:

* project_essay_1: "Describe your students: What makes your students special? Specific details about their background, your neighborhood, and your school are all helpful."
* project_essay_2: "About your project: How will these materials make a difference in your students' learning and improve their school lives?"

For all projects with project_submitted_datetime of 2016-05-17 and later, the values of project_essay_3 and project_essay_4 will be NaN.

max(pre_cleaned_proposals.loc[(pre_cleaned_proposals['datetime']>datetime.datetime(2016, 5, 17))                          
                          & (pd.notnull(pre_cleaned_proposals['essay_3'])), 'datetime'])

max(pre_cleaned_proposals.loc[(pre_cleaned_proposals['datetime']>datetime.datetime(2016, 5, 17))                          
                          & (pd.notnull(pre_cleaned_proposals['essay_4'])), 'datetime'])

pre_cleaned_proposals.loc[(pre_cleaned_proposals['datetime']>datetime.datetime(2016, 5, 17))                          
                          & (pre_cleaned_proposals['datetime']<datetime.datetime(2016, 5, 17, 12))
                             , ['essay_3', 'essay_4', 'datetime']].sort_values('datetime')

pre_cleaned_proposals['essay_3'].isnull().sum()

pre_cleaned_proposals['essay_4'].isnull().sum()

##### Combine questions 1 and 2, for proposals before May 17, 2016

pre_cleaned_proposals.loc[(pre_cleaned_proposals['datetime']<datetime.datetime(2016, 5, 17, 9, 39)),
                        'essay1'] = pre_cleaned_proposals['essay_1'] + ' ' + pre_cleaned_proposals['essay_2']

pre_cleaned_proposals['essay1'].isnull().sum()

pre_cleaned_proposals.loc[(pre_cleaned_proposals['datetime']>datetime.datetime(2016, 5, 17, 9, 39)),
                        'essay1'] = pre_cleaned_proposals['essay_1']

# check
pre_cleaned_proposals['essay1'].isnull().sum()

##### Combine questions 3 and 4, for proposals before May 17, 2016

pre_cleaned_proposals.loc[(pre_cleaned_proposals['datetime']<datetime.datetime(2016, 5, 17, 9, 39)),
                        'essay2'] = pre_cleaned_proposals['essay_3'] + ' ' + pre_cleaned_proposals['essay_4']

pre_cleaned_proposals['essay2'].isnull().sum()

pre_cleaned_proposals.loc[(pre_cleaned_proposals['datetime']>datetime.datetime(2016, 5, 17, 9, 39)),
                        'essay2'] = pre_cleaned_proposals['essay_2'] 

pre_cleaned_proposals['essay2'].isnull().sum()

###### Clean all text columns

text_columns = ['essay1',
                'essay2',
                'resource_summary',
                'total_description']

for column in text_columns:
    print(column)
    new_col_name = 'c_' + column
    pre_cleaned_proposals[new_col_name] = pre_cleaned_proposals[column].apply(process_text)

pre_cleaned_proposals.to_pickle('pre_cleaned_proposals.pkl')
