==> Capstone/4_FinalPrep.ipynb <==
# Capstone Project: Udacity Machine Learning Engineer
## Hans Fricke, 05-18-2018
## Final data prep

# import packages
import pandas as pd
import os
import numpy as np
import pprint

# change working directory
os.chdir("/Users/hansfricke/Dropbox/ND ML Engineer/Capstone")

#### Read all data sets

proposals = pd.read_pickle('Data/pre_cleaned_proposals.pkl')

topics = pd.read_pickle('Data/nmf_topics.pkl')

components = pd.read_pickle('Data/pca_components.pkl')

#### Join data sets

# check that all have same amount of rows
print(len(proposals))
print(len(topics))
print(len(components))

# get list of column names
proposals_cols = list(proposals.columns)
pprint.pprint(proposals_cols)

# get list of column names
topics_cols = list(topics.columns)
pprint.pprint(topics_cols)

# get list of column names
components_cols = list(components.columns)
pprint.pprint(list(components_cols))

##### Merge data sets

final_data = pd.merge(proposals, topics, on='id')

final_data = pd.merge(final_data, components, on='id')

#### Select columns used in analysis

remove_cols = ['teacher_id',
                 'teacher_prefix',
                 'state',
                 'datetime',
                 'grade_category',
                 'subject_categories',
                 'subject_subcategories',
                 'title',
                 'essay_1',
                 'essay_2',
                 'essay_3',
                 'essay_4',
                 'resource_summary', 
                 'essay1',
                 'essay2', 
                 'total_description',
                 'c_essay1',
                 'c_essay2',
                 'c_resource_summary',
                 'c_total_description']
selected_cols = [col for col in proposals_cols if col not in remove_cols]

topics_cols.remove('id')
components_cols.remove('id')

selected_cols.extend(topics_cols)

selected_cols.extend(components_cols)

pprint.pprint(selected_cols)

final_data = final_data[selected_cols]

final_data.to_pickle('Data/final_data.pkl')
